♻ Refactor:멘토링 수정사항 리팩토리 - 제미나이 코드리뷰 나머지 반영: RoleProp를 공용 types파일로 이동
