+++ src/app/(service)/[teamid]/types/team.ts
@@ -2,7 +2,9 @@
 // 팀 / 그룹 관련 타입
 // ========================
 
-import type { ModalKey, RoleProps } from './common';
+import { RoleProps } from '@/app/types';
+
+import type { ModalKey } from './common';
 import type { MemberChipsProps } from './member';
 import type { TaskList } from './task';
